dist: Autogenerate AUTHORS file
Per discussion at the Summer 2017 developers meeting, generate
the AUTHORS list at make dist time, rather than trying to
keep it up to date and merge on the branches by hand.  While
most of the data is generated from git, the organization list
was maintained by hand.  The general feeling at the meeting was
that the organization list was not adding value and there were
concrete cases where it involved much chasing by the RMs, so
it has been removed.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>